refactor: ページ移動でSwitchを初期化
